refactor: split main into probe/retry/sidecar modules
