+++ src/main.rs
@@ -23,13 +23,16 @@ use std::{
     os::raw::c_void,
     path::{Path, PathBuf},
     ptr,
-    sync::atomic::{AtomicI64, AtomicU64, Ordering},
+    sync::atomic::{AtomicI64, Ordering},
 };
 
 mod config;
 mod devices;
 mod gpu;
 mod logging;
+mod main_probe;
+mod main_retry;
+mod main_sidecar;
 #[cfg(test)]
 mod main_tests;
 mod plex;
@@ -40,6 +43,14 @@ mod throttle;
 use gpu::{
     acquire_hw_device, find_hw_encoder, gather_probe_json, print_probe, print_probe_codecs, HwAccel,
 };
+use main_probe::{gather_streams_info_json, print_streams_info};
+use main_retry::{
+    cleanup_partial_output, handle_hw_encoder_init_error, handle_hw_profile_mismatch,
+    retry_with_software_encoder, select_primary_video_stream_index,
+};
+use main_sidecar::post_process_ocr_subtitles;
+#[cfg(test)]
+use main_sidecar::{sidecar_path_for_track, write_ocr_srt_sidecars};
 use throttle::acquire_slot;
 
 fn describe_bitrate(bitrate: Option<i64>) -> String {
@@ -4372,423 +4383,6 @@ fn convert_video_file(
     Ok(ConversionOutcome { h264_verification })
 }
 
-fn cleanup_partial_output(path: &CStr) {
-    let output_path = PathBuf::from(path.to_string_lossy().into_owned());
-    if output_path.exists() {
-        if let Err(remove_err) = fs::remove_file(&output_path) {
-            warn!(
-                "Failed to remove incompatible output '{}': {}",
-                output_path.display(),
-                remove_err
-            );
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn retry_with_software_encoder(
-    input_file: &CStr,
-    output_file: &CStr,
-    sub_mode: SubMode,
-    target_video_codec: ffi::AVCodecID,
-    target_audio_codec: ffi::AVCodecID,
-    h264_constraints: Option<(H264Profile, H264Level)>,
-    min_fps: u32,
-    min_resolution: Resolution,
-    quality_limits: &QualityLimits,
-    uv_policy: UnsupportedVideoPolicy,
-    primary_video_stream_index: Option<usize>,
-    primary_video_criteria: PrimaryVideoCriteria,
-    requested_video_quality: VideoQuality,
-    requested_audio_quality: AudioQuality,
-    skip_codec_check: bool,
-) -> Result<ConversionOutcome, anyhow::Error> {
-    convert_video_file(
-        input_file,
-        output_file,
-        sub_mode,
-        target_video_codec,
-        target_audio_codec,
-        h264_constraints,
-        min_fps,
-        min_resolution,
-        quality_limits,
-        uv_policy,
-        primary_video_stream_index,
-        primary_video_criteria,
-        requested_video_quality,
-        requested_audio_quality,
-        skip_codec_check,
-        HwAccel::None,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn handle_hw_profile_mismatch(
-    mismatch: HwProfileLevelMismatch,
-    args: &Args,
-    input_file: &CStr,
-    output_file: &CStr,
-    sub_mode: SubMode,
-    target_video_codec: ffi::AVCodecID,
-    target_audio_codec: ffi::AVCodecID,
-    h264_constraints: Option<(H264Profile, H264Level)>,
-    min_fps: u32,
-    min_resolution: Resolution,
-    quality_limits: &QualityLimits,
-) -> Result<ConversionOutcome, anyhow::Error> {
-    if args.hw_accel == HwAccel::None || !mismatch.used_hw_encoder {
-        return Err(anyhow!(mismatch));
-    }
-
-    let actual_profile = mismatch
-        .actual_profile
-        .map(|p| format!("{:?}", p))
-        .unwrap_or_else(|| "unknown".to_string());
-    let actual_level = mismatch
-        .actual_level
-        .map(|l| l.ffmpeg_name().to_string())
-        .unwrap_or_else(|| "unknown".to_string());
-    warn!(
-        "Hardware encoder {} produced H.264 profile {} level {} for '{}' (expected profile {:?} level {:?}); retrying with software encoder (libx264)",
-        mismatch.encoder,
-        actual_profile,
-        actual_level,
-        mismatch.output_path,
-        mismatch.expected_profile,
-        mismatch.expected_level
-    );
-    cleanup_partial_output(output_file);
-    retry_with_software_encoder(
-        input_file,
-        output_file,
-        sub_mode,
-        target_video_codec,
-        target_audio_codec,
-        h264_constraints,
-        min_fps,
-        min_resolution,
-        quality_limits,
-        args.unsupported_video_policy,
-        args.primary_video_stream_index,
-        args.primary_video_criteria,
-        args.video_quality,
-        args.audio_quality,
-        args.skip_codec_check,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn handle_hw_encoder_init_error(
-    init_error: HwEncoderInitError,
-    args: &Args,
-    input_file: &CStr,
-    output_file: &CStr,
-    sub_mode: SubMode,
-    target_video_codec: ffi::AVCodecID,
-    target_audio_codec: ffi::AVCodecID,
-    h264_constraints: Option<(H264Profile, H264Level)>,
-    min_fps: u32,
-    min_resolution: Resolution,
-    quality_limits: &QualityLimits,
-) -> Result<ConversionOutcome, anyhow::Error> {
-    if args.hw_accel == HwAccel::None {
-        return Err(anyhow!(init_error));
-    }
-
-    warn!(
-        "Hardware encoder {} failed to initialize ({}); retrying with software encoder",
-        init_error.encoder, init_error.message
-    );
-    cleanup_partial_output(output_file);
-    retry_with_software_encoder(
-        input_file,
-        output_file,
-        sub_mode,
-        target_video_codec,
-        target_audio_codec,
-        h264_constraints,
-        min_fps,
-        min_resolution,
-        quality_limits,
-        args.unsupported_video_policy,
-        args.primary_video_stream_index,
-        args.primary_video_criteria,
-        args.video_quality,
-        args.audio_quality,
-        args.skip_codec_check,
-    )
-}
-
-fn select_primary_video_stream_index(
-    input_ctx: &AVFormatContextInput,
-    override_index: Option<usize>,
-    criteria: PrimaryVideoCriteria,
-) -> Result<usize> {
-    if let Some(idx) = override_index {
-        let streams = input_ctx.streams();
-        if idx >= streams.len() {
-            bail!(
-                "--primary-video-stream-index={} out of range (streams: {})",
-                idx,
-                streams.len()
-            );
-        }
-        let st = &streams[idx];
-        if st.codecpar().codec_type != ffi::AVMEDIA_TYPE_VIDEO {
-            bail!("--primary-video-stream-index={} is not a video stream", idx);
-        }
-        return Ok(idx);
-    }
-
-    let mut best_idx: Option<usize> = None;
-    let mut best_score: u128 = 0;
-    for st in input_ctx.streams() {
-        if st.codecpar().codec_type != ffi::AVMEDIA_TYPE_VIDEO {
-            continue;
-        }
-        let cp = st.codecpar();
-        let (w, h) = (cp.width as u64, cp.height as u64);
-        let area = w.saturating_mul(h);
-        let br = if cp.bit_rate > 0 {
-            cp.bit_rate as u64
-        } else {
-            0
-        };
-        let fps_milli: u64 = st
-            .guess_framerate()
-            .map(|tb| {
-                let num = tb.num as i128;
-                let den = if tb.den == 0 { 1 } else { tb.den } as i128;
-                let v = (num * 1000) / den;
-                if v < 0 {
-                    0
-                } else {
-                    v as u64
-                }
-            })
-            .unwrap_or(0);
-        let score: u128 = match criteria {
-            PrimaryVideoCriteria::Resolution => ((area as u128) << 40) + (br as u128),
-            PrimaryVideoCriteria::Bitrate => ((br as u128) << 40) + (area as u128),
-            PrimaryVideoCriteria::Fps => ((fps_milli as u128) << 56) + (area as u128),
-        };
-        if best_idx.is_none() || score > best_score {
-            best_idx = Some(st.index as usize);
-            best_score = score;
-        }
-    }
-    best_idx.ok_or_else(|| anyhow!("No video streams found in input"))
-}
-
-fn print_streams_info(input_file: &CStr, filter: StreamsFilter) -> Result<()> {
-    let ictx = AVFormatContextInput::open(input_file)?;
-    println!("Input: {}", input_file.to_string_lossy());
-    let duration_us = ictx.duration;
-    if duration_us > 0 {
-        println!("Duration: {} ms", duration_us / 1000);
-    }
-    for st in ictx.streams() {
-        let idx = st.index;
-        let tb = st.time_base;
-        let cp = st.codecpar();
-        let ctype = cp.codec_type;
-        let kind_matches = match filter {
-            StreamsFilter::All => true,
-            StreamsFilter::Video => ctype == ffi::AVMEDIA_TYPE_VIDEO,
-            StreamsFilter::Audio => ctype == ffi::AVMEDIA_TYPE_AUDIO,
-            StreamsFilter::Subtitle => ctype == ffi::AVMEDIA_TYPE_SUBTITLE,
-        };
-        if !kind_matches {
-            continue;
-        }
-        let cname = unsafe { std::ffi::CStr::from_ptr(ffi::avcodec_get_name(cp.codec_id)) };
-        let (stream_id, disp_default, disp_forced, disp_hi, disp_vi) = unsafe {
-            let s_ptr = st.as_ptr();
-            let id = (*s_ptr).id;
-            let d = (*s_ptr).disposition;
-            (
-                id,
-                (d & ffi::AV_DISPOSITION_DEFAULT as i32) != 0,
-                (d & ffi::AV_DISPOSITION_FORCED as i32) != 0,
-                (d & ffi::AV_DISPOSITION_HEARING_IMPAIRED as i32) != 0,
-                (d & ffi::AV_DISPOSITION_VISUAL_IMPAIRED as i32) != 0,
-            )
-        };
-        print!(
-            "[stream {} id={}] type={:?} codec={} ",
-            idx,
-            stream_id,
-            ctype,
-            cname.to_string_lossy()
-        );
-        match ctype {
-            ffi::AVMEDIA_TYPE_VIDEO => {
-                let (w, h) = (cp.width, cp.height);
-                let fps = st
-                    .guess_framerate()
-                    .map(|r| format!("{}/{}", r.num, r.den))
-                    .unwrap_or_else(|| "?".into());
-                println!(
-                    "res={}x{} fps={} bitrate={} time_base={}/{}",
-                    w, h, fps, cp.bit_rate, tb.num, tb.den
-                );
-            }
-            ffi::AVMEDIA_TYPE_AUDIO => {
-                let ch = cp.ch_layout.nb_channels;
-                println!(
-                    "channels={} sample_rate={} bitrate={} time_base={}/{}",
-                    ch, cp.sample_rate, cp.bit_rate, tb.num, tb.den
-                );
-            }
-            ffi::AVMEDIA_TYPE_SUBTITLE => {
-                println!("subtitle time_base={}/{}", tb.num, tb.den);
-            }
-            _ => {
-                println!("time_base={}/{}", tb.num, tb.den);
-            }
-        }
-        println!(
-            "  disposition: default={} forced={} hearing_impaired={} visual_impaired={}",
-            disp_default, disp_forced, disp_hi, disp_vi
-        );
-    }
-    Ok(())
-}
-
-#[derive(Serialize)]
-struct JsonStreamInfo {
-    index: i32,
-    stream_id: i32,
-    kind: String,
-    codec: String,
-    time_base: (i32, i32),
-    #[serde(skip_serializing_if = "Option::is_none")]
-    width: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    height: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    fps: Option<(i32, i32)>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    bitrate: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    channels: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    sample_rate: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    language: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    title: Option<String>,
-    disposition: JsonDisposition,
-}
-
-#[derive(Serialize)]
-struct JsonProbe {
-    input: String,
-    duration_ms: Option<i64>,
-    streams: Vec<JsonStreamInfo>,
-}
-
-#[derive(Serialize)]
-struct JsonDisposition {
-    default: bool,
-    forced: bool,
-    hearing_impaired: bool,
-    visual_impaired: bool,
-}
-
-fn gather_streams_info_json(input_file: &CStr, filter: StreamsFilter) -> Result<JsonProbe> {
-    let ictx = AVFormatContextInput::open(input_file)?;
-    let duration_us = ictx.duration;
-    let mut out: Vec<JsonStreamInfo> = Vec::new();
-    for st in ictx.streams() {
-        let idx = st.index;
-        let tb = st.time_base;
-        let cp = st.codecpar();
-        let ctype = cp.codec_type;
-        let kind = match ctype {
-            ffi::AVMEDIA_TYPE_VIDEO => "video",
-            ffi::AVMEDIA_TYPE_AUDIO => "audio",
-            ffi::AVMEDIA_TYPE_SUBTITLE => "subtitle",
-            ffi::AVMEDIA_TYPE_ATTACHMENT => "attachment",
-            ffi::AVMEDIA_TYPE_DATA => "data",
-            _ => "other",
-        }
-        .to_string();
-        let kind_matches = match filter {
-            StreamsFilter::All => true,
-            StreamsFilter::Video => ctype == ffi::AVMEDIA_TYPE_VIDEO,
-            StreamsFilter::Audio => ctype == ffi::AVMEDIA_TYPE_AUDIO,
-            StreamsFilter::Subtitle => ctype == ffi::AVMEDIA_TYPE_SUBTITLE,
-        };
-        if !kind_matches {
-            continue;
-        }
-        let cname = unsafe { std::ffi::CStr::from_ptr(ffi::avcodec_get_name(cp.codec_id)) };
-        let (stream_id, disp_default, disp_forced, disp_hi, disp_vi) = unsafe {
-            let s_ptr = st.as_ptr();
-            let id = (*s_ptr).id;
-            let d = (*s_ptr).disposition;
-            (
-                id,
-                (d & ffi::AV_DISPOSITION_DEFAULT as i32) != 0,
-                (d & ffi::AV_DISPOSITION_FORCED as i32) != 0,
-                (d & ffi::AV_DISPOSITION_HEARING_IMPAIRED as i32) != 0,
-                (d & ffi::AV_DISPOSITION_VISUAL_IMPAIRED as i32) != 0,
-            )
-        };
-        let (mut width, mut height, mut fps, mut channels, mut sample_rate) =
-            (None, None, None, None, None);
-        match ctype {
-            ffi::AVMEDIA_TYPE_VIDEO => {
-                width = Some(cp.width);
-                height = Some(cp.height);
-                fps = st.guess_framerate().map(|r| (r.num, r.den));
-            }
-            ffi::AVMEDIA_TYPE_AUDIO => {
-                channels = Some(cp.ch_layout.nb_channels);
-                sample_rate = Some(cp.sample_rate);
-            }
-            _ => {}
-        }
-        let (language, title): (Option<String>, Option<String>) = (None, None);
-        out.push(JsonStreamInfo {
-            index: idx,
-            stream_id,
-            kind,
-            codec: cname.to_string_lossy().into_owned(),
-            time_base: (tb.num, tb.den),
-            width,
-            height,
-            fps,
-            bitrate: if cp.bit_rate > 0 {
-                Some(cp.bit_rate)
-            } else {
-                None
-            },
-            channels,
-            sample_rate,
-            language,
-            title,
-            disposition: JsonDisposition {
-                default: disp_default,
-                forced: disp_forced,
-                hearing_impaired: disp_hi,
-                visual_impaired: disp_vi,
-            },
-        });
-    }
-    Ok(JsonProbe {
-        input: input_file.to_string_lossy().into_owned(),
-        duration_ms: if duration_us > 0 {
-            Some(duration_us / 1000)
-        } else {
-            None
-        },
-        streams: out,
-    })
-}
-
 fn main() -> Result<()> {
     if env::var_os("RUST_LOG").is_none() {
         env::set_var("RUST_LOG", "info");
@@ -5382,158 +4976,4 @@ fn run_conversion(
         }
         (None, Err(err)) => Err(err),
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn post_process_ocr_subtitles(
-    input_file: &CStr,
-    mux_source_file: &CStr,
-    output_file: &CStr,
-    sub_mode: SubMode,
-    default_ocr_language: Option<&str>,
-    ocr_engine: OcrEngine,
-    ocr_format: OcrFormat,
-    ocr_external_command: Option<&str>,
-    ocr_write_srt_sidecar: bool,
-) -> Result<()> {
-    if matches!(sub_mode, SubMode::Skip) {
-        return Ok(());
-    }
-
-    let ocr_work_dir = OcrWorkDir::create()?;
-
-    let tracks = subtitle_ocr::convert_bitmap_subtitles(
-        input_file,
-        ocr_work_dir.path(),
-        sub_mode,
-        default_ocr_language,
-        ocr_engine,
-        ocr_format,
-        ocr_external_command,
-    )?;
-    subtitle_ocr::mux_text_tracks_from(mux_source_file, output_file, &tracks)?;
-    write_ocr_srt_sidecars(output_file, &tracks, ocr_write_srt_sidecar)?;
-
-    Ok(())
-}
-
-struct OcrWorkDir {
-    path: PathBuf,
-}
-
-static OCR_WORK_DIR_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-impl OcrWorkDir {
-    fn create() -> Result<Self> {
-        let mut path = std::env::temp_dir();
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_millis())
-            .unwrap_or(0);
-        let nonce = OCR_WORK_DIR_COUNTER.fetch_add(1, Ordering::Relaxed);
-        path.push(format!(
-            "direct-play-nice-ocr-{}-{}-{}",
-            std::process::id(),
-            now,
-            nonce
-        ));
-        fs::create_dir(&path)
-            .with_context(|| format!("creating OCR temporary directory '{}'", path.display()))?;
-        Ok(Self { path })
-    }
-
-    fn path(&self) -> &Path {
-        &self.path
-    }
-}
-
-impl Drop for OcrWorkDir {
-    fn drop(&mut self) {
-        if let Err(err) = fs::remove_dir_all(&self.path) {
-            debug!(
-                "Failed to clean OCR temporary directory '{}': {}",
-                self.path.display(),
-                err
-            );
-        }
-    }
-}
-
-fn sanitize_sidecar_language(language: &str) -> String {
-    let normalized: String = language
-        .trim()
-        .to_ascii_lowercase()
-        .chars()
-        .filter(|ch| ch.is_ascii_alphanumeric() || *ch == '-' || *ch == '_')
-        .collect();
-    if normalized.is_empty() {
-        "und".to_string()
-    } else {
-        normalized
-    }
-}
-
-fn sidecar_path_for_track(
-    output_path: &Path,
-    language: &str,
-    language_occurrence: usize,
-) -> PathBuf {
-    let parent = output_path.parent().unwrap_or_else(|| Path::new("."));
-    let stem = output_path
-        .file_stem()
-        .and_then(|value| value.to_str())
-        .unwrap_or("output");
-    let suffix = if language_occurrence <= 1 {
-        format!("{stem}.{language}.srt")
-    } else {
-        format!("{stem}.{language}.{language_occurrence}.srt")
-    };
-    parent.join(suffix)
-}
-
-fn write_ocr_srt_sidecars(
-    output_file: &CStr,
-    tracks: &[subtitle_ocr::OcrSubtitleTrack],
-    enabled: bool,
-) -> Result<()> {
-    if !enabled || tracks.is_empty() {
-        return Ok(());
-    }
-
-    let output_path = PathBuf::from(output_file.to_string_lossy().into_owned());
-    let mut language_counts: HashMap<String, usize> = HashMap::new();
-
-    for track in tracks {
-        if !matches!(track.format, OcrFormat::Srt) {
-            warn!(
-                "Skipping OCR sidecar write for '{}' because format is {:?} (SRT required).",
-                track.subtitle_path.display(),
-                track.format
-            );
-            continue;
-        }
-
-        let language = sanitize_sidecar_language(&track.language);
-        let occurrence = {
-            let count = language_counts.entry(language.clone()).or_insert(0);
-            *count += 1;
-            *count
-        };
-        let sidecar_path = sidecar_path_for_track(&output_path, &language, occurrence);
-        fs::copy(&track.subtitle_path, &sidecar_path).with_context(|| {
-            format!(
-                "writing OCR sidecar '{}' from '{}'",
-                sidecar_path.display(),
-                track.subtitle_path.display()
-            )
-        })?;
-        info!(
-            "Wrote OCR sidecar subtitle '{}' (language={}, source='{}').",
-            sidecar_path.display(),
-            language,
-            track.subtitle_path.display()
-        );
-    }
-
-    Ok(())
 }
